requirements.txt
streamlit
opencv-python-headless
deepface
tf-keras
numpy
pillow